Add a container library.
Added Support for strings and dynamic arrays. Which are quite similar to the
std::vector and std::string but don't suffer from dll issues. They are
templated but currently no instantations are exported to the outside in the
public headers.

BaseString and DArray are integrated in libeye making the c dll basically
obsolete. There is however still an issue with the BaseString<char>
implementation

Added the instantations of the Templated classes. and the new import header
for libeye since the previous was based on the old C intermediate layer.

Fixed bug in build that overwrites CXXFLAGS

Completed the buildin types for libeye. Also added tests for string. For now
the assumption is that the tests of the BaseString reasonalbly cover the test
for DArray, since it uses it heavely.

Fixed bug in writing a csv type log that didn't place the line terminators
where they should be.

Add new container library.

For shared libraries it is not a wise idea to export the C++ standard
library, since it is not binary compatible with other C++ compilers.
Therefore this library adds its own Container library
to avoid.