Add support for pkg-config and change log dir to eyelog
Add support for pkg-config by a .pc file. Changed log dir
to eyelog. The all future packages can include the stuff in the
log as <eyelog/EyeLog.h> which will import pretty much everything.
The two programs in the bin directory also follow this as of now.
This is so that those two programs can serve as an example.